+++ pkg/service/handlers/observability/alert.go
@@ -1824,10 +1824,10 @@ func (p *AlertRuleProcessor) GetK8sAlertCfg(ctx context.Context) (map[string]K8s
 
 func (h *ObservabilityHandler) syncAlertRulesWithTimeout(ctx context.Context, alertrules []*models.AlertRule, timeout time.Duration) (status map[string]bool, isTimeout bool) {
 	// sync alert rules
-	status = map[string]bool{}
+	tmp := sync.Map{}
 	wg := &sync.WaitGroup{}
 	for _, v := range alertrules {
-		status[v.FullName()] = false
+		tmp.Store(v.FullName(), false)
 		wg.Add(1)
 		go func(alertrule *models.AlertRule) {
 			defer wg.Done()
@@ -1837,9 +1837,14 @@ func (h *ObservabilityHandler) syncAlertRulesWithTimeout(ctx context.Context, al
 				log.Warnf("%s alert rule: %s sync failed", alertrule.AlertType, alertrule.FullName())
 				return
 			}
-			status[alertrule.FullName()] = true
+			tmp.Store(alertrule.FullName(), false)
 		}(v)
 	}
+	status = map[string]bool{}
+	tmp.Range(func(key, value any) bool {
+		status[key.(string)] = value.(bool)
+		return true
+	})
 	isTimeout = utils.WaitGroupWithTimeout(wg, timeout)
 	if isTimeout {
 		log.Warnf("Timed out waiting for wait group")