+++ scripts/offline-plugins/main.go
@@ -15,9 +15,13 @@
 package main
 
 import (
+	"bufio"
 	"bytes"
 	"context"
+	"errors"
+	"flag"
 	"fmt"
+	"io"
 	"log"
 	"os"
 	"os/signal"
@@ -25,6 +29,7 @@ import (
 	"strings"
 	"syscall"
 
+	"golang.org/x/exp/slices"
 	"helm.sh/helm/v3/pkg/repo"
 	"k8s.io/client-go/kubernetes/scheme"
 	plugins "kubegems.io/kubegems/pkg/apis/plugins"
@@ -35,27 +40,95 @@ import (
 	"kubegems.io/kubegems/pkg/installer/utils"
 )
 
+var _ = pluginv1beta1.AddToScheme(scheme.Scheme)
+
 func main() {
 	ctx, cancel := signal.NotifyContext(context.Background(), os.Interrupt, os.Kill, syscall.SIGTERM)
 	defer cancel()
 
-	// download latest charts
-	if err := DownloadLatestCharts(ctx, plugins.KubegemsChartsRepoURL, "bin/plugins", "latest"); err != nil {
+	from := flag.String("from", "plugins.txt", "offline plugin versions from")
+	to := flag.String("to", "bin/plugins", "offline plugins to")
+	flag.Parse()
+
+	if err := Run(ctx, *from, *to); err != nil {
 		fmt.Printf("Error: %s\n", err.Error())
 		os.Exit(1)
 	}
 }
 
-var _ = pluginv1beta1.AddToScheme(scheme.Scheme)
-
-func DownloadLatestCharts(ctx context.Context, repoaddr string, into string, kubegemsVersion string) error {
-	applier := bundle.NewDefaultApply(nil, nil, &bundle.Options{CacheDir: into})
-	pluginrepo := pluginmanager.Repository{Address: repoaddr}
-	if err := pluginrepo.RefreshRepoIndex(ctx); err != nil {
+func Run(ctx context.Context, from, to string) error {
+	offline := NewOffline(plugins.KubegemsChartsRepoURL)
+	if err := offline.ReadFromFile(ctx, from); err != nil {
 		return err
 	}
-	kubegemsExample := pluginmanager.PluginVersion{Name: "kubegems", Version: kubegemsVersion}
-	for name, versions := range pluginrepo.Plugins {
+	if err := offline.Download(ctx, to); err != nil {
+		return err
+	}
+	return nil
+}
+
+func NewOffline(repoaddr string) *Offline {
+	return &Offline{
+		repo:           pluginmanager.Repository{Address: repoaddr},
+		offlineplugins: map[string]*pluginmanager.PluginVersion{},
+	}
+}
+
+type Offline struct {
+	repo           pluginmanager.Repository
+	offlineplugins map[string]*pluginmanager.PluginVersion
+}
+
+func (o *Offline) ReadFromFile(ctx context.Context, filename string) error {
+	caches, err := ReadPluginFile(filename)
+	if err != nil {
+		if !errors.Is(err, os.ErrNotExist) {
+			return err
+		}
+		// ignore not found
+	}
+	if err := o.repo.RefreshRepoIndex(ctx); err != nil {
+		return err
+	}
+	if len(caches) == 0 {
+		o.offlineplugins = readLatestVersions(ctx, &o.repo)
+		// writeback
+		caches := []OfflinePlugin{}
+		for _, v := range o.offlineplugins {
+			log.Printf("found latest: %s-%s", v.Name, v.Version)
+			caches = append(caches, OfflinePlugin{Name: v.Name, Version: v.Version})
+		}
+		if err := WritePluginFile(filename, caches); err != nil {
+			return err
+		}
+	} else {
+		o.offlineplugins = readCachedVersions(ctx, &o.repo, caches)
+	}
+	return nil
+}
+
+func readCachedVersions(ctx context.Context, repo *pluginmanager.Repository, cached []OfflinePlugin) map[string]*pluginmanager.PluginVersion {
+	ret := map[string]*pluginmanager.PluginVersion{}
+	for _, cache := range cached {
+		found := false
+		for _, cv := range repo.Plugins[cache.Name] {
+			if cv.Version == cache.Version {
+				ret[cache.Name] = &cv
+				found = true
+				log.Printf("found: %s-%s", cache.Name, cache.Version)
+				break
+			}
+		}
+		if !found {
+			log.Printf("version not found: %s-%s", cache.Name, cache.Version)
+		}
+	}
+	return ret
+}
+
+func readLatestVersions(ctx context.Context, repo *pluginmanager.Repository) map[string]*pluginmanager.PluginVersion {
+	ret := map[string]*pluginmanager.PluginVersion{}
+	for name, versions := range repo.Plugins {
 		// do not download kubegems charts,it exists locally.
 		if strings.HasPrefix(name, "kubegems") {
 			continue
@@ -63,20 +136,24 @@ func DownloadLatestCharts(ctx context.Context, repoaddr string, into string, kub
 		var cacheVersion *pluginmanager.PluginVersion
 		// find latest version match kubegems
 		for _, item := range versions {
-			if err := pluginmanager.CheckDependecy(item.Requirements, kubegemsExample); err != nil {
-				log.Printf("ignore plugin [%s-%s] on dependencis not match: %s", item.Name, item.Version, err.Error())
-				continue
-			} else {
-				cacheVersion = &item
-				break
-			}
+			cacheVersion = &item
+			break
 		}
 		if cacheVersion == nil {
 			log.Printf("no matched version to cache on plugin %s", name)
 			continue
 		}
-		log.Printf("download %s-%s from %s", cacheVersion.Name, cacheVersion.Version, repoaddr)
-		manifest, err := applier.Template(ctx, cacheVersion.ToPlugin())
+		ret[name] = cacheVersion
+	}
+	return ret
+}
+
+func (o *Offline) Download(ctx context.Context, basedir string) error {
+	applier := bundle.NewDefaultApply(nil, nil, &bundle.Options{CacheDir: basedir})
+	for _, pv := range o.offlineplugins {
+		plugin := pv.ToPlugin()
+		log.Printf("download %s-%s from %s", plugin.Name, plugin.Spec.Version, plugin.Spec.Kind)
+		manifest, err := applier.Template(ctx, plugin)
 		if err != nil {
 			log.Printf("on template: %v", err)
 			return err
@@ -100,7 +177,7 @@ func DownloadLatestCharts(ctx context.Context, repoaddr string, into string, kub
 		}
 	}
 	// build index
-	indexpath := bundle.PerRepoCacheDir(repoaddr, into)
+	indexpath := bundle.PerRepoCacheDir(o.repo.Address, basedir)
 	log.Printf("generating helm repo index.yaml under %s", indexpath)
 	i, err := repo.IndexDirectory(indexpath, "")
 	if err != nil {
@@ -108,4 +185,64 @@ func DownloadLatestCharts(ctx context.Context, repoaddr string, into string, kub
 	}
 	i.SortEntries()
 	return i.WriteFile(filepath.Join(indexpath, "index.yaml"), helm.DefaultFileMode)
+}
+
+func ReadPluginFile(filename string) ([]OfflinePlugin, error) {
+	data, err := os.ReadFile(filename)
+	if err != nil {
+		return nil, err
+	}
+	return ParseContent(data), nil
+}
+
+func WritePluginFile(filename string, list []OfflinePlugin) error {
+	data := bytes.NewBuffer(nil)
+	slices.SortFunc(list, func(a, b OfflinePlugin) bool {
+		return strings.Compare(a.Name, b.Name) == -1
+	})
+	for _, val := range list {
+		data.WriteString(val.Name + " " + val.Version + "\n")
+	}
+	return os.WriteFile(filename, data.Bytes(), helm.DefaultFileMode)
+}
+
+type OfflinePlugin struct {
+	Name    string
+	Version string
+}
+
+func ParseContent(data []byte) []OfflinePlugin {
+	scan := bufio.NewReader(bytes.NewReader(data))
+	ret := []OfflinePlugin{}
+	for {
+		line, _, err := scan.ReadLine()
+		if err == io.EOF {
+			break
+		}
+		line = bytes.TrimLeft(line, " ")
+		if len(line) == 0 {
+			continue
+		}
+		if line[0] == '#' || line[0] == ';' {
+			continue
+		}
+		fields := bytes.Fields(line)
+		if len(fields) == 0 {
+			continue
+		}
+		switch len(fields) {
+		case 0:
+			continue
+		case 1:
+			ret = append(ret, OfflinePlugin{
+				Name: string(fields[0]),
+			})
+		default:
+			ret = append(ret, OfflinePlugin{
+				Name:    string(fields[0]),
+				Version: string(fields[1]),
+			})
+		}
+	}
+	return ret
 }